A start towards specifying current and next TAP
This is a slightly incomplete BNF expression of the TAP format.

I don't know of a clever way to express the concept of indentation in
BNF.  The intent is to specify that child tests shall be indented 4
spaces, and lack the "TAP version 14" header, and start with a name
comment, but in all other ways from a grammar point of view, should be
just like any other "<test-set>" block.

Furthermore, there's a lot of *semantics* that go into "valid TAP", of
course.  This is just the start of defining the syntax.

I also don't know a reasonable way to say "comments can go anywhere",
without making the BNF very unweidly.